# Slices: Optimize SpanEnumerator