This fruit was hanging very low.
To let  JIT enregister all the fields I've "inlined" Span into Span enumerator.
It makes it almost equal (in codegen and perf) to for loop over a local copy of the Slice.

### Benchmarks
[Code](https://gist.github.com/omariom/5c54b85eb747f5b46b58)
**Results**: 
BenchmarkDotNet-Dev=v0.8.0+
OS=Microsoft Windows NT 6.9200.0
Processor=Intel(R) Core(TM) i7-3537U CPU @ 00GHz, ProcessorCount=4
HostCLR=MS.NET 0.30319.42000, Arch=64-bit  [RyuJIT]
Type=BenchmarksEnumerator  Mode=Throughput  Platform=X64  Jit=HostJit  .NET=HostFramework  toolchain=Classic  Runtime=Clr  Warmup=2  Target=2  

                Method | IntParam |       AvrTime |      StdDev |        op/s |
---------------------- |--------- |-------------- |------------ |------------ |
   OLD Slice enumerate |    **8** |    68.0764 ns |  29.3606 ns | 16,195,6553 |
       Slice enumerate |        8 |    29.0776 ns |   0.4373 ns | 34,394,6107  |
   Slice for over copy |        8 |      32159 ns |   0.2540 ns | 28,397,0178 |
             Slice for |        8 |    27.6080 ns |    19134 ns | 42,408,7618 |
      String enumerate |        8 |    18.3388 ns |   7.7780 ns | 59,918,5641 |
  Char array enumerate |        8 |    16.1519 ns |     2713 ns | 64,155,5481 |
   OLD Slice enumerate |   **16** |    91.5581 ns |  35.8995 ns | 11,831,5045 |
       Slice enumerate |       16 |    39.8605 ns |    19452 ns | 26,266,9371 |
   Slice for over copy |       16 |    56.9489 ns |    10284 ns | 18,109,0284 |
             Slice for |       16 |    57.6788 ns |  27.5462 ns | 19,569,0678 |
      String enumerate |       16 |    38.5514 ns |   0.4500 ns | 25,941,1492 |
  Char array enumerate |       16 |      35175 ns |   9.5032 ns | 32,124,5663 |
   OLD Slice enumerate |   **32** |   204.0937 ns |  10.3003 ns |  4,905,9572 |
       Slice enumerate |       32 |      71196 ns |    26078 ns | 14,428,2325 |
   Slice for over copy |       32 |   117.2667 ns |    29400 ns |  8,679,4805 |
             Slice for |       32 |   107.7834 ns |    25141 ns |  9,466,4555 |
      String enumerate |       32 |      75001 ns |    15395 ns | 13,495,2506 |
  Char array enumerate |       32 |    56.9351 ns |   8.3369 ns | 17,754,1907 |
   OLD Slice enumerate |   **64** |   286.2561 ns |  45.5209 ns |  3,538,1108 |
       Slice enumerate |       64 |   118.1565 ns |  16.4553 ns |  8,546,2306 |
   Slice for over copy |       64 |   156.1408 ns |  36.4911 ns |  6,584,2901 |
             Slice for |       64 |   178.5550 ns |  37.2670 ns |  5,725,2147 |
      String enumerate |       64 |   130.9740 ns |   0.7461 ns |  7,635,2233 |
  Char array enumerate |       64 |     114199 ns |    24743 ns |  8,816,6178 |
   OLD Slice enumerate |  **256** | 1,059.9809 ns | 105.1558 ns |    948,0785 |
       Slice enumerate |      256 |   500.7829 ns |     8893 ns |  1,996,8873 |
   Slice for over copy |      256 |     459193 ns |    67733 ns |  2,224,3039 |
             Slice for |      256 |   607.4890 ns |     9870 ns |  1,646,1402 |
      String enumerate |      256 |   368.9185 ns |     3167 ns |  2,710,6432 |
  Char array enumerate |      256 |     369700 ns |   6.3075 ns |  2,732,8700 |
   OLD Slice enumerate | **1024** | 4,533.5800 ns | 1,122.23 ns |    227,5477 |
       Slice enumerate |     1024 |   1,997753 ns |     3428 ns |    501,5614 |
   Slice for over copy |     1024 |   1,883222 ns |   961522 ns |    611,8997 |
             Slice for |     1024 | 2,230.4255 ns |    61104 ns |    448,5130 |
      String enumerate |     1024 |   1,191613 ns | 530.7023 ns |    928,2171 |
  Char array enumerate |     1024 | 1,039.5923 ns |   345932 ns |  1,017,1467 |
   OLD Slice enumerate | **2048** | 8,006.6383 ns | 1,407.92 ns |    126,8576 |
       Slice enumerate |     2048 |   4,287701 ns | 980.0816 ns |    239,5920 |
   Slice for over copy |     2048 |   5,047217 ns | 226.6893 ns |    198,4273 |
             Slice for |     2048 | 4,367.0600 ns | 199.9807 ns |    229,2273 |
      String enumerate |     2048 | 1,976.9005 ns |   544800 ns |    525,7084 |
  Char array enumerate |     2048 |   2,634682 ns |   510673 ns |    387,3206 |