@KrzysztofCwalina 2016-02-09
looks great. thanks!

cc: @sokket, @livarcocc